pyproject.toml
[project]
name = "large-passenger-aircraft"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "flask-jwt-extended>=4.7.1",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "marshmallow>=3.26.1",
    "mysqlclient>=2.2.7",
    "pydantic>=2.11.3",
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
    "python-dotenv>=1.1.0",
    "sqlalchemy-serializer>=1.4.22",
]
